+++ src/migrations/2020_02_10_054032_add_nexus_compatibility.php
@@ -18,11 +18,30 @@ class AddNexusCompatibility extends Migration
                 ->nullable()
                 ->default(null);
         });
+        Schema::table('categories', function (Blueprint $table) {
+            $table->integer('nexus_id')
+                ->nullable()
+                ->default(null);
+        });
         Schema::table('torrents', function (Blueprint $table) {
             $table->integer('nexus_id')
                 ->nullable()
                 ->default(null);
             $table->integer('nexus_user_id')
+                ->nullable()
+                ->default(null);
+            $table->integer('nexus_category_id')
+                ->nullable()
+                ->default(null);
+        });
+        Schema::table('comments', function (Blueprint $table) {
+            $table->integer('nexus_id')
+                ->nullable()
+                ->default(null);
+            $table->integer('nexus_user_id')
+                ->nullable()
+                ->default(null);
+            $table->integer('nexus_torrent_id')
                 ->nullable()
                 ->default(null);
         });
@@ -40,6 +59,9 @@ class AddNexusCompatibility extends Migration
             $table->integer('nexus_parent_id')
                 ->nullable()
                 ->default(null);
+            $table->string('nexus_permission')
+                ->nullable()
+                ->default(null);
         });
         Schema::table('topics', function (Blueprint $table) {
             $table->integer('nexus_id')
@@ -60,6 +82,14 @@ class AddNexusCompatibility extends Migration
                 ->nullable()
                 ->default(null);
             $table->integer('nexus_parent_id')
+                ->nullable()
+                ->default(null);
+        });
+        Schema::table('messages', function (Blueprint $table) {
+            $table->integer('nexus_id')
+                ->nullable()
+                ->default(null);
+            $table->integer('nexus_user_id')
                 ->nullable()
                 ->default(null);
         });
@@ -75,13 +105,23 @@ class AddNexusCompatibility extends Migration
         Schema::table('users', function (Blueprint $table) {
             $table->dropColumn('nexus_id');
         });
+        Schema::table('categories', function (Blueprint $table) {
+            $table->dropColumn('nexus_id');
+        });
         Schema::table('torrents', function (Blueprint $table) {
             $table->dropColumn('nexus_id');
             $table->dropColumn('nexus_user_id');
+            $table->dropColumn('nexus_category_id');
+        });
+        Schema::table('comments', function (Blueprint $table) {
+            $table->dropColumn('nexus_id');
+            $table->dropColumn('nexus_user_id');
+            $table->dropColumn('nexus_torrent_id');
         });
         Schema::table('forums', function (Blueprint $table) {
             $table->dropColumn('nexus_id');
             $table->dropColumn('nexus_parent_id');
+            $table->dropColumn('nexus_permission');
         });
         Schema::table('topics', function (Blueprint $table) {
             $table->dropColumn('nexus_id');
@@ -92,6 +132,10 @@ class AddNexusCompatibility extends Migration
             $table->dropColumn('nexus_id');
             $table->dropColumn('nexus_user_id');
             $table->dropColumn('nexus_parent_id');
+        });
+        Schema::table('messages', function (Blueprint $table) {
+            $table->dropColumn('nexus_id');
+            $table->dropColumn('nexus_user_id');
         });
     }
 }